+++ core/src/test/scala/peloton/actors/CollectorActor.scala
@@ -24,26 +24,28 @@ object CollectorActor:
   final case class GetResponse(words: List[String])
   given CanAsk[Get, GetResponse] = canAsk
 
+  // TODO: this talk about different solutions is basically documentation and as such it should be moved to an example file
+
   // Solution 1 (preferred): functional
   def handlerFn(words: List[String]): Behavior[State, Message] =
     (state, message, context) => 
       message match
         case Add(word) => 
-          context.respond(AddResponse(word)) >> 
+          context.reply(AddResponse(word)) >> 
           handlerFn(words :+ word).pure
 
         case Get() => 
-          context.respond(GetResponse(words))
+          context.reply(GetResponse(words))
 
   // Solution 2: with a modifiable state
   val handler: Behavior[State, Message] =
     (state, message, context) => message match
       case Add(word) => 
         context.setState(State(state.words :+ word)) >> 
-        context.respond(AddResponse(word))
+        context.reply(AddResponse(word))
 
       case Get() => 
-        context.respond(GetResponse(state.words))
+        context.reply(GetResponse(state.words))
 
   def spawn(name: String = "CollectorActor")(using actorSystem: ActorSystem) = 
     actorSystem.spawn[State, Message](